feat(raycast,render): add DDA helpers and per-ray timing instrumentation
Add DDA helper functions for raycasting (`calculate_hit`, `perform_dda`) and
side setter helpers (`set_side_capitalist`, `set_side_communist`). Introduce a
render timing API (`t_rendertiming`) and functions to record/reset per-ray and
frame timings (`rendertiming_add_ray`, `rendertiming_reset`,
`rendertiming_set_all_rays`, `rendertiming_set_render`) and embed timing state
into `t_renderctx` to enable ray-level performance instrumentation.

Core changes:
- Add: `src/graphics/raycast/{calculate_hit.c,perform_dda.c,set_side_capitalist.c,set_side_communist.c}`
- Modify: `src/graphics/raycast/raycast_dda.c` (use `t_renderctx`)
- Add: `src/graphics/render/rendertiming_{add_ray,reset,set_all_rays,set_render}.c`
- Update: `include/graphics/raycast.h`, `include/graphics/render.h` for new prototypes and types
- Small build/editor updates: `Makefile`, `.vscode/extensions.json`

Impact:
- Developer: Some prototypes changed (notably `raycast_dda` signature and raycast prototypes in headers) — update callers accordingly.
- Performance: Adds instrumentation only (no behavior change). Useful for profiling and future optimizations.
- Tests: None added in this batch; recommend unit tests for raycasting correctness and timing aggregation.

BREAKING CHANGE: `raycast_dda` and related prototypes in `include/graphics/raycast.h` were updated; callers must be recompiled/adjusted to match the new signatures.

+++ src/graphics/render/render_update.c
@@ -6,7 +6,7 @@
 /*   By: lfiorell <lfiorell@student.42nice.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/05 12:48:16 by lfiorell          #+#    #+#             */
-/*   Updated: 2026/01/05 13:36:18 by lfiorell         ###   ########.fr       */
+/*   Updated: 2026/01/06 15:02:29 by lfiorell         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,8 +34,8 @@ static bool	can_move(t_renderctx *ctx, t_vec2 *new_pos)
 
 	new_cell_x = (int)(new_pos->x);
 	new_cell_y = (int)(new_pos->y);
-	if (new_cell_x < 0 || new_cell_x >= ctx->map->width || new_cell_y < 0
-		|| new_cell_y >= ctx->map->height)
+	if (new_cell_x < 0 || new_cell_x >= (int)ctx->map->width || new_cell_y < 0
+		|| new_cell_y >= (int)ctx->map->height)
 		return (false);
 	return (true);
 }
